SaveAs / LoadFrom / Compression * Added saveManger for Trie. * Add loadFromFile and saveAs (closes #31). * Used XWRT to compress XML file (closes #30). * Added LoadThread. * Added open program from command line * Notes:     * LoadFromFile clears the log first, I'm thinking of merging the tries instead of clearing.     * .gitignore don't allow .exe files, you should include xwert32.exe 3.4.     * We're using static dicitionary compression. wrt-eng contains nearly all english words. We should include only words may exist in the XML file for better compression.
