+++ ext.reg
@@ -1,0 +1,14 @@
+Windows Registry Editor Version 5.00
+
+[HKEY_CLASSES_ROOT\.plcl]
+@="PolynomialCalculator"
+
+[HKEY_CLASSES_ROOT\PolynomialCalculator]
+@="Polynomial Calculator History"
+
+[HKEY_CLASSES_ROOT\PolynomialCalculator\shell]
+
+[HKEY_CLASSES_ROOT\PolynomialCalculator\shell\open]
+
+[HKEY_CLASSES_ROOT\PolynomialCalculator\shell\open\command]
+@="\"Our path\" \"%1\""
